fix scan_xlong, scan_xlonglong and scan_8long

+++ scan/scan_8long.c
@@ -2,7 +2,7 @@
 
 unsigned int scan_8long(const char *src,unsigned long *dest) {
   register const char *tmp=src;
-  register int l=0;
+  register unsigned long l=0;
   register unsigned char c;
   while ((c=*tmp-'0')<8) {
     l=l*8+c;
