+++ scan/scan_xlong.c
@@ -2,7 +2,7 @@
 
 unsigned int scan_xlong(const char *src,unsigned long *dest) {
   register const char *tmp=src;
-  register int l=0;
+  register unsigned long l=0;
   register unsigned char c;
   while ((c=scan_fromhex(*tmp))<16) {
     l=(l<<4)+c;
